+++ examples/karma/karma.conf.js
@@ -24,6 +24,7 @@ module.exports = config => {
         },
         kelonioReporter: {
             inferBrowsers: true,
+            printReportAtEnd: true,
         },
     });
 };